make auto align feature to hoc function
