+++ stories/index.js
@@ -1,7 +1,6 @@
 import './BadgeStories';
 import './BreadCrumbsStories';
 import './ButtonStories';
-import './DropdownStories';
 import './DropdownButtonStories';
 import './ButtonGroupStories';
 import './DatepickerStories';
